clip-service/requirements.txt
open-clip-torch
pillow
fastapi
uvicorn
requests
pydantic

